add song_classifier: classify song ids: base class for distillation
