+++ utako/analyzer/__init__.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # import utako.analyzer.chainer_chart_analyzer
+import utako.analyzer.song_classifier
 import utako.analyzer.song_auto_encoder
 import utako.analyzer.sae_tag_comparator
 # import utako.analyzer.sklearn_chart_analyzer
